feat: use api route instead of server action
